requirements.txt
fastapi==0.95.1
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
pydantic==1.10.7
python-multipart==0.0.6
typing_extensions==4.5.0
uvicorn==0.22.0
